% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_genes_all_clusters.R
\name{assign_genes_all_clusters}
\alias{assign_genes_all_clusters}
\title{Classify genes into clusters}
\usage{
assign_genes_all_clusters(
  exprMatr,
  cell_clusters,
  H0_cells = NULL,
  N_cores = 1,
  backend = "doParallel"
)
}
\arguments{
\item{exprMatr}{The input expression matrix.}

\item{cell_clusters}{The clusters obtained by cell cluster.}

\item{H0_cells}{The specific list of reference cells, default "NULL".}

\item{N_cores}{The cores used, default "1".}

\item{backend}{set the operating mode, default "doParallel"}
}
\description{
Classify genes into clusters
}
