% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdjustedRandIndex.R
\name{CalculateARI}
\alias{CalculateARI}
\title{Calculate ARI of true_lab pre_lab}
\usage{
CalculateARI(true_lab, pre_lab)
}
\arguments{
\item{true_lab}{The original cell type information.}

\item{pre_lab}{The predicted cell type information.}
}
\description{
Calculate ARI of true_lab pre_lab
}
