% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Compute_CCI.R
\name{Compute_CCI}
\alias{Compute_CCI}
\title{Classify genes into clusters}
\usage{
Compute_CCI(
  data = data,
  Label = Label,
  cell_group = cell_group,
  LR_pairs_DB = LR_pairs_DB,
  TF_reg_DB = TF_TG,
  R_TF_association = TF_PPR,
  LR_db = LR_db,
  N_cores = 2,
  DEmethod = "wilcoxon",
  n_bootstrap = 4,
  backend = "doParallel",
  other_inter_scores = NULL,
  cell_reference = NULL,
  Org = Org,
  use.type = "median",
  probs = 0.9,
  method = "weighted"
)
}
\arguments{
\item{data}{The input expression matrix.}

\item{cell_group}{The clusters obtained by cell cluster.}
}
\description{
Classify genes into clusters
}
