% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_score_S_ligand_receptor1.R
\name{compute_score_S_ligand_receptor1}
\alias{compute_score_S_ligand_receptor1}
\title{Compute the pvalue of target genes in all cluster}
\usage{
compute_score_S_ligand_receptor1(
  exprMatr,
  cells,
  ligands,
  receptors,
  H0_genes,
  verbose = TRUE,
  debug_result = FALSE
)
}
\arguments{
\item{exprMatr}{The input expression matrix.}

\item{cells}{The input cell_id.}

\item{ligands}{The input lig_id.}

\item{receptors}{The input rec_id.}

\item{H0_genes}{The input array of gene names.}
}
\description{
Compute the pvalue of target genes in all cluster
}
