% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Compute_Sintra.R
\name{Compute_Sintra}
\alias{Compute_Sintra}
\title{Compute the intercellular interaction}
\usage{
Compute_Sintra(
  gene_expr = gene_expr,
  cell_group = cell_group,
  LR_pairs_DB = LR_pairs_DB,
  TF_reg_DB = TF_TG,
  R_TF_association = TF_PPR,
  LR_db = row_index_final,
  N_cores = 2,
  DEmethod = "wilcoxon",
  n_bootstrap = 4,
  backend = "doParallel",
  other_inter_scores = NULL,
  cell_reference = NULL
)
}
\arguments{
\item{gene_expr}{The input expression matrix.}

\item{cell_group}{The input cell_group.}

\item{LR_pairs_DB}{The input ligand-receptor interactions.}

\item{TF_reg_DB}{The transcription factor-target regulatory interactions.}

\item{R_TF_association}{The input PPR_scores.}

\item{LR_db}{The input ligand-receptor interactions.}
}
\description{
Compute the intercellular interaction
}
