% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DrjccLG.R
\name{DrjccLG}
\alias{DrjccLG}
\title{The main function of DcjComm}
\usage{
DrjccLG(params)
}
\arguments{
\item{params}{The input params inludes input matrix X, cell-cell similarity W, dimension k1 and k2.}
}
\description{
The main function of DcjComm
}
