% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ligand_receptor_all_cluster_score.R
\name{Ligand_receptor_all_cluster_score}
\alias{Ligand_receptor_all_cluster_score}
\title{Compute the ligand_receptor communications in all cluster}
\usage{
Ligand_receptor_all_cluster_score(
  exprMatr,
  cell_cluster,
  ligands,
  receptors,
  H0_genes = NULL,
  verbose = FALSE,
  debug_result = FALSE,
  bootstrap = FALSE,
  bootstrap_size = 1
)
}
\arguments{
\item{exprMatr}{The input expression matrix.}

\item{ligands}{The input target genes.}

\item{receptors}{The target genes score.}

\item{cell_clusters}{The clusters obtained by cell cluster.}
}
\description{
Compute the ligand_receptor communications in all cluster
}
