% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moduleNodesSelection.R
\name{moduleNodesSelection}
\alias{moduleNodesSelection}
\title{Assign genes to different clusters}
\usage{
moduleNodesSelection(moduleparams)
}
\arguments{
\item{moduleparams}{The input params.}
}
\description{
Assign genes to different clusters
}
