% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFbasedjoint.R
\name{NMFbased}
\alias{NMFbased}
\title{The main function of DcjComm}
\usage{
NMFbased(params)
}
\arguments{
\item{params}{The input params inludes input matrix X, cell-cell similarity W, dimension k1 and k2.}
}
\description{
The main function of DcjComm
}
