% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pvalue_target_genes.R
\name{Pvalue_target_genes}
\alias{Pvalue_target_genes}
\title{Compute the cell-cell interaction}
\usage{
Pvalue_target_genes(
  exprMatr,
  ref_exprMatr,
  target_genes,
  unreliable_genes,
  DEmethod = "MAST",
  N_cores = 1
)
}
\arguments{
\item{gene_expr}{The input expression matrix.}

\item{TF_PPR}{The signal transductions from receptor to transcription factors.}

\item{TF_TG}{The transcription factor-target regulatory interactions.}

\item{S_intra}{The calculated results of intracellular score.}

\item{row_index_final}{The ligand-receptor association.}

\item{Org}{The input species information.}
}
\description{
Compute the cell-cell interaction
}
