% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Receptor_pathway_score.R
\name{Receptor_pathway_score}
\alias{Receptor_pathway_score}
\title{Claculate the association score of receptor and pathways}
\usage{
Receptor_pathway_score(TFscore, tfweights, method)
}
\arguments{
\item{TFscore}{The calculated tf score.}

\item{tfweights}{The tfweights obtained obove.}

\item{method}{The computation method of score, eg "sum", "mean".}

\item{N_cores}{The cores used, default "1".}

\item{backend}{set the operating mode, default "doParallel"}
}
\description{
Claculate the association score of receptor and pathways
}
