% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S_intra_all_clusters.R
\name{S_intra_all_clusters}
\alias{S_intra_all_clusters}
\title{Compute the cell-cell interaction}
\usage{
S_intra_all_clusters(
  R_pathway_TF_weight,
  TFscore,
  cell_cluster,
  method = c("sum", "mean")
)
}
\arguments{
\item{R_pathway_TF_weight}{The input PPR_scores.}

\item{TFscore}{The obtained TF_activity.}

\item{cell_cluster}{The above constructed cell_group.}
}
\description{
Compute the cell-cell interaction
}
