% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{scSeqComm_chorddiag_cardinality}
\alias{scSeqComm_chorddiag_cardinality}
\title{Plot a chord diagram showing the amount of cell-cell communication}
\usage{
scSeqComm_chorddiag_cardinality(
  data,
  var1 = "cluster_R",
  var2 = "cluster_L",
  group_fontsize = 12,
  margin = 90,
  var1_label = "Receptor-expressing cell types",
  var2_label = "Ligand-expressing cell types"
)
}
\arguments{
\item{data}{A dataframe containing the results of signaling analysis.}

\item{var1}{A character string representing the first category group. It must be a subset of \code{colnames(data)}.}

\item{var2}{A character string representing the second category group. It must be a subset of \code{colnames(data)}.}

\item{group_fontsize}{Numeric font size in pixels for the group name labels.}

\item{margin}{Numeric margin in pixels between the outer diagram radius and the edge of the display.}

\item{var1_label}{A character string to be used for category var1 label}

\item{var2_label}{A character string to be used for category var2 label}
}
\value{
The function returns an image in the plot window of Rstudio (Viewer)
}
\description{
Create an interactive bipartite chord diagram showing number of observation among two categories of groups (var1-var2).
}
