% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{scSeqComm_heatmap}
\alias{scSeqComm_heatmap}
\title{Summary plot (heatmap)}
\usage{
scSeqComm_heatmap(
  data,
  x,
  y,
  fill = "S_intra",
  size = "S_inter",
  title,
  xlab,
  ylab = "",
  fill_lab = "S_intra",
  size_lab = "S_inter",
  limit_fill = c(0, 1),
  limit_size = c(0, 1),
  breaks_size = c(0.2, 0.5, 0.8),
  facet_grid_x = NULL,
  facet_grid_y = NULL,
  annotation_GO = NULL,
  cutoff = 0.05,
  topGO = 5,
  GO_ncol = 1
)
}
\arguments{
\item{data}{A data frame containing results of signaling analysis.}

\item{x}{A character string describing the variable to be mapped on x axis. It must be a subset of \code{colnames(data)}.}

\item{y}{A character string describing the variable to be mapped on y axis. It must be a subset of \code{colnames(data)}.}

\item{fill}{A character string describing the variable to be mapped in color scale. It must be a subset of \code{colnames(data)}.}

\item{size}{a character string describing the variable to be mapped in point size scale. It must be a subset of \code{colnames(data)}.}

\item{title}{A character string to be used as title of plot.}

\item{xlab}{A character string to be used as title of x axis.}

\item{ylab}{A character string to be used as title of y axis.}

\item{fill_lab}{A character string to be used as label for fill scale.}

\item{size_lab}{A character string to be used as label for size scale.}

\item{limit_fill}{A numeric vector of length two describing the fill scale limits.}

\item{limit_size}{A numeric vector of length two describing the size scale limits.}

\item{breaks_size}{A numeric vector of positions of breaks.}

\item{facet_grid_x}{A variable defining faceting group on columns dimension.}

\item{facet_grid_y}{A variable defining faceting group on row dimension.}

\item{annotation_GO}{A data.frame or a list of data.frame with results of enrichment analysis.}

\item{cutoff}{A numeric value defining the significance value for GO terms.}

\item{topGO}{A numeric value defining the maximum number of terms to be visualized.}

\item{GO_ncol}{A numeric value defining the number of columns on which show the GO terms table.}
}
\value{
The function returns the image in the plot window of Rstudio
}
\description{
The function creates a heatmap based on specified parameters.
}
