% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{scSeqComm_heatmap_cardinality}
\alias{scSeqComm_heatmap_cardinality}
\title{Plot an heatmap showing the amount of cell-cell communication}
\usage{
scSeqComm_heatmap_cardinality(
  data,
  y = "cluster_L",
  x = "cluster_R",
  limit_fill = NULL,
  cell_types = NULL,
  ylab = "Ligand-expressing cell types",
  xlab = "Receptor-expressing cell types",
  title
)
}
\arguments{
\item{data}{A dataframe containing results of signaling analysis.}

\item{y}{A character string describing the variable to be plotted on y axis. It must be a subset of \code{colnames(data)}.}

\item{x}{A character string describing the variable to be plotted on x axis. It must be a subset of \code{colnames(data)}.}

\item{limit_fill}{A numeric vector of length two providing limits of the scale. Use NULL to refer to the existing minimum or maximum in \code{data}.}

\item{cell_types}{A character vector containing the cell types to be plotted in the heatmap.}

\item{ylab}{A character string to be used as title of y axis.}

\item{xlab}{A character string to be used as title of x axis.}

\item{title}{A character string to be used as title of plot.}
}
\value{
The function returns an image in the plot window of Rstudio
}
\description{
Creates an heatmap showing the number of observations among two categories of groups (x-y)
}
