% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{scSeqComm_plot_LR_pair}
\alias{scSeqComm_plot_LR_pair}
\title{Plot intercellular and intracellular scores of a given ligand-receptor pair across cell clusters (heatmap)}
\usage{
scSeqComm_plot_LR_pair(data, title, selected_LR_pair = NULL)
}
\arguments{
\item{data}{A data frame containing results of signaling analysis.}

\item{title}{A character string to be used as title of plot.}

\item{selected_LR_pair}{String or array of strings containing ligand-receptor pairs to plot.}
}
\value{
A plot (or a list of plots)
}
\description{
Function to plot intercellular (mapped in point size scale) and intracellular (mapped in color scale) scores of a given ligand-receptor pair across cell clusters.
If input data contains more than one ligand-receptor pair, a plot will be generated for each ligand-receptor pair.
Grey color is associated to NA values of intracellular score.
}
