% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{scSeqComm_plot_scores}
\alias{scSeqComm_plot_scores}
\title{Plot intercellular and intracellular scores (heatmap)}
\usage{
scSeqComm_plot_scores(
  data,
  title,
  facet_grid_x = "cluster_R",
  facet_grid_y = "receptor",
  annotation_GO = NULL,
  cutoff = 0.05,
  topGO = 5,
  GO_ncol = 1
)
}
\arguments{
\item{data}{A data frame containing results of signaling analysis.}

\item{title}{A character string to be used as title of plot.}

\item{facet_grid_x}{A variable defining faceting group on columns dimension.}

\item{facet_grid_y}{A variable defining faceting group on row dimension.}

\item{annotation_GO}{A data.frame or a list of data.frame with results of enrichment analysis.}

\item{cutoff}{A numeric value defining the significance value for GO terms.}

\item{topGO}{A numeric value defining the maximum number of terms to be visualized.}

\item{GO_ncol}{A numeric value defining the number of columns on which show the GO terms table.}
}
\value{
Plot showing intracellular and intercellular scores
}
\description{
Function to plot the intercellular (mapped in point size scale) and intracellular (mapped in color scale) scores of the given input data with cluster pairs on x axis and ligand-receptor pairs on y axis.
Grey color is associated to NA values of intracellular score.
}
