% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tfactors_score.R
\name{Tfactors_score}
\alias{Tfactors_score}
\title{Compute the tfactors score}
\usage{
Tfactors_score(regulatoryDB, genes_score, cutoff)
}
\arguments{
\item{regulatoryDB}{The input regulatoryDB.}

\item{genes_score}{The TGpvalues obtained above.}

\item{cutoff}{Select genes after Bonferroni correction according to cutoff, default "0.05.}
}
\description{
Compute the tfactors score
}
