% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Compute_CCI.R
\name{Compute_CCI}
\alias{Compute_CCI}
\title{Compute the cell-cell interaction}
\usage{
Compute_CCI(
  gene_expr = gene_expr,
  TF_PPR = TF_PPR,
  TF_TG = TF_TG,
  S_intra = S_intra,
  row_index_final = row_index_final,
  Org = Org,
  use.type = "median",
  probs = 0.9,
  method = "weighted"
)
}
\arguments{
\item{gene_expr}{The input expression matrix.}

\item{TF_PPR}{The signal transductions from receptor to transcription factors.}

\item{TF_TG}{The transcription factor-target regulatory interactions.}

\item{S_intra}{The calculated results of intracellular score.}

\item{row_index_final}{The ligand-receptor association.}

\item{Org}{The input species information.}
}
\description{
Compute the cell-cell interaction
}
